# [FSDP2] Added eager fast-path for fp32->bf16 param cast