@awgu 2024-08-15
cc: @weifengpy depending on how many weights use fp8, the fp8 casting CPU overhead, even with the precompute, might become a problem